test(ci): extract managed image workflow types
