+++ test/managed-image-publication-workflow.test.ts
@@ -16,69 +16,12 @@ import {
   runManagedImagePromotion,
   runPublicationBarrier,
 } from "./helpers/managed-image-publication-barrier";
-
-type Step = {
-  env?: Record<string, unknown>;
-  id?: string;
-  if?: string;
-  name?: string;
-  run?: string;
-  uses?: string;
-  with?: Record<string, unknown>;
-  "working-directory"?: string;
-};
-
-type MatrixEntry = {
-  agent?: string;
-  arch?: string;
-  artifact_platform?: string;
-  base_alias?: string;
-  base_dockerfile?: string;
-  base_image?: string;
-  base_repository?: string;
-  display_name?: string;
-  dockerfile?: string;
-  image?: string;
-  repository?: string;
-  platform?: string;
-  required_binary?: string;
-  runner?: string;
-};
-
-type Job = {
-  env?: Record<string, unknown>;
-  if?: string;
-  needs?: string | string[];
-  permissions?: Record<string, string>;
-  "runs-on"?: string;
-  steps?: Step[];
-  strategy?: {
-    "fail-fast"?: boolean;
-    matrix?: { include?: MatrixEntry[] };
-  };
-  "timeout-minutes"?: number;
-  uses?: string;
-};
-
-type Workflow = {
-  concurrency?: {
-    "cancel-in-progress"?: string | boolean;
-    group?: string;
-  };
-  env?: Record<string, string>;
-  jobs?: Record<string, Job>;
-  on?: {
-    pull_request?: {
-      branches?: string[];
-      paths?: string[];
-    };
-    push?: {
-      paths?: string[];
-    };
-    workflow_call?: unknown;
-  };
-  permissions?: Record<string, string>;
-};
+import type {
+  Job,
+  MatrixEntry,
+  Step,
+  Workflow,
+} from "./helpers/managed-image-publication-workflow-types";
 
 const repoRoot = path.resolve(import.meta.dirname, "..");
 const fullShaAction = /^[^@]+@[0-9a-f]{40}$/iu;
@@ -1496,5 +1439,34 @@ fi
         "langchain-deepagents-code": expect.any(Object),
       },
     });
+  });
+
+  it("retains exact platform and aggregate cohort contracts for ninety days (#7744)", () => {
+    const promoter = managedPromoter(readWorkflow("managed-images.yaml"));
+    const uploads = (promoter.steps ?? [])
+      .filter((candidate) => candidate.uses?.startsWith("actions/upload-artifact@"))
+      .map((candidate) => candidate.with);
+
+    expect(uploads).toEqual([
+      {
+        name: "managed-image-cohort-${{ github.run_id }}-${{ github.run_attempt }}",
+        path: "${{ runner.temp }}/managed-image-contracts/cohort.json",
+        "if-no-files-found": "error",
+        "retention-days": 90,
+      },
+      ...publicationAgents.flatMap((agent) =>
+        publicationPlatforms.map((platform) => {
+          const artifactPlatform = platform.replaceAll("/", "-");
+          return {
+            name:
+              "managed-image-${{ github.run_id }}-${{ github.run_attempt }}-" +
+              `${agent}-${artifactPlatform}`,
+            path: `\${{ runner.temp }}/managed-image-contracts/${agent}/${artifactPlatform}/contract.json`,
+            "if-no-files-found": "error",
+            "retention-days": 90,
+          };
+        }),
+      ),
+    ]);
   });
 });
